Add support for querying SauceLabs for capacity
Similar to how we query Selenium Grids to find out if they have
available capacity before sending sessions to them, let's do the same
for SauceLabs.

Added a new function to the SauceLabs handler for querying the
concurrency endpoint of the SauceLabs API. This function returns True if
there is available capacity, False otherwise. This function is then
called before we construct the SauceLabs URL for the test. Similar to
how we keep trying to send a session to the Selenium Grid until the
timeout is reached, we keep checking SauceLabs for available capacity
until the timeout is exceeded. Both of these timeouts are currently
controlled by the same configuration option.

Additionally, added unit tests covering the new feature. Introduced
`requests_mock` as a new test dependency to make testing against the
SauceLabs API a less annoying task.