Removed DynamoDB Dependency
Removed the dependency on DynamoDB for storing state information.
Instead, the session id is used to store this information. This is
achieved by hashing the URL of the grid that a session is using, and
appending that to the end of the session id that is generated by
Selenium. This hash is then parsed out by Amplium on subsequent requests
and used to lookup the Selenium Grid that the session should be
forwarded to.

In addition to the above change, took advantage of the time to rewrite
the backend a little bit to have more cleanly deliniated lines between
the different backend classes. For example, a lot of the internal logic
that was hanging off of proxy.py and zookeeper.py has been refactored
into a new utility class, grid_handler.py