Handle Zookeeper Nodes Disappearing Bug
Apparently, it was possible for some kind of bad state to develop where
the Zookeeper helper's list of nodes would contain an old node that no
longer exists but the watcher wouldn't be called to update it. Now if we
encounter that situation, we can catch the error and turn around and
forcibly update the list of nodes.