Add Connection Retry Policy
The default settings make it so that Amplium continually retries
connections to Zookeeper. This means that sometimes Amplium can get
stuck into an infinite loop where it continually tries to connect to
Zookeeper. Adding this policy so that we have an upper bound of retries,
and the process will die and Apache will bring it back if we end up in
some strange state.